chore: rename match > IMatch
