+++ app/frontend/src/game.ts
@@ -1,7 +1,7 @@
 import { updatePaddlePositions, setupInputListeners } from "./input.js";
 import { draw } from "./draw.js";
 import { IGameState } from "./types/IGameState.js";
-import { Match } from "./types/match.js"
+import { Match } from "./types/IMatch.js"
 import NewGame from "./views/NewGame.js";
 
 export function renderGame() {
@@ -21,27 +21,27 @@ export function renderGame() {
 }
 
 function initGameState(canvas: HTMLCanvasElement, nickname1: string, nickname2: string): IGameState {
-  return {
-      player1: nickname1,
-      player2: nickname2,
-      player1Score: 0,
-      player2Score: 0,
-      winningScore: 3,
-      ballX: canvas.width / 2,
-      ballY: canvas.height / 2,
-      ballSpeedX: 7,
-      ballSpeedY: 7,
-      paddle1Y: canvas.height / 2 - 40,
-      paddle2Y: canvas.height / 2 - 40,
-      paddleHeight: 80,
-      paddleWidth: 10,
-      paddleSpeed: 6,
-      gameStarted: false,
-      gameOver: false
-  };
+	return {
+		player1: nickname1,
+		player2: nickname2,
+		player1Score: 0,
+		player2Score: 0,
+		winningScore: 3,
+		ballX: canvas.width / 2,
+		ballY: canvas.height / 2,
+		ballSpeedX: 7,
+		ballSpeedY: 7,
+		paddle1Y: canvas.height / 2 - 40,
+		paddle2Y: canvas.height / 2 - 40,
+		paddleHeight: 80,
+		paddleWidth: 10,
+		paddleSpeed: 6,
+		gameStarted: false,
+		gameOver: false
+	};
 }
 
-function startGame(canvas: HTMLCanvasElement,  ctx: CanvasRenderingContext2D, gameState: IGameState) {
+function startGame(canvas: HTMLCanvasElement, ctx: CanvasRenderingContext2D, gameState: IGameState) {
 	gameState.gameOver = false;
 	gameState.gameStarted = true;
 	document.getElementById("register-form")?.remove();
