+++ app/frontend/src/types/IUserStats.ts
